feat: server improvement

+++ src/modules/categories/categories.controller.ts
@@ -3,6 +3,7 @@ import {
   Controller,
   Delete,
   Get,
+  HttpStatus,
   Param,
   Post,
   Query,
@@ -10,7 +11,9 @@ import {
 } from '@nestjs/common';
 import {
   ApiBearerAuth,
+  ApiBody,
   ApiOperation,
+  ApiParam,
   ApiQuery,
   ApiResponse,
   ApiTags,
@@ -20,46 +23,62 @@ import { Category } from './category.entity';
 import { CreateCategoryDto } from './dto/create.category.dto';
 import { CategoriesService } from './categories.service';
 import { UtilsService } from '../common/utils/utils.service';
+import { CurrentUser } from '../users/decorators/user.decorator';
+import { isAdmin, isLandlord, UserPayload } from '../auth/dto/user.payload';
+import { Public } from '../common/decorators/public.decorator';
+import { AccessRightsException } from '../common/exceoptions/access.rights.exception';
 
-@Controller('types')
-@ApiTags('Виды отелей 📃')
+@Controller('categories')
+@ApiTags('Категории отелей 📃')
 export class CategoriesController {
   constructor(private readonly categoriesService: CategoriesService) {}
 
-  @ApiOperation({ summary: 'Создание новой категории отеля' })
   @ApiBearerAuth()
   @UseGuards(AuthGuard('jwt'))
+  @ApiOperation({ summary: 'Создание новой категории отеля' })
   @ApiResponse({
-    status: 201,
+    status: HttpStatus.CREATED,
     description: 'Категория отеля успешно создана',
     type: Category,
   })
   @ApiResponse({
-    status: 400,
+    status: HttpStatus.BAD_REQUEST,
     description: 'Некорректные данные категории отеля',
   })
+  @ApiBody({
+    type: CreateCategoryDto,
+    description: 'Данные для создания новой категории отеля',
+  })
   @Post()
-  async create(@Body() addedCategory: CreateCategoryDto): Promise<Category> {
-    const isNameExist = await this.categoriesService.isNameExist(
-      addedCategory.name,
-    );
+  async create(
+    @CurrentUser() user: UserPayload,
+    @Body() addedCategory: CreateCategoryDto,
+  ): Promise<Category> {
+    if (isAdmin(user) || isLandlord(user)) {
+      const isNameExist = await this.categoriesService.isNameExist(
+        addedCategory.name,
+      );
 
-    if (!isNameExist) {
-      return await this.categoriesService.create(addedCategory);
+      if (!isNameExist) {
+        return await this.categoriesService.create(addedCategory);
+      }
+    } else {
+      throw new AccessRightsException();
     }
   }
 
   @ApiOperation({ summary: 'Получение списка всех категорий отеля' })
   @ApiResponse({
-    status: 200,
+    status: HttpStatus.OK,
     description: 'Список всех категорий',
-    type: [Category],
+    type: Category,
   })
   @ApiQuery({
     name: 'name',
     description: 'Название, закодированное в base64',
     required: false,
   })
+  @Public()
   @Get()
   async getAll(@Query('name') name?: string): Promise<Category[]> {
     return await this.categoriesService.findAll(
@@ -69,10 +88,16 @@ export class CategoriesController {
 
   @ApiOperation({ summary: 'Получить категорию отеля по ID' })
   @ApiResponse({
-    status: 200,
+    status: HttpStatus.OK,
     description: 'Информация о категории',
     type: Category,
   })
+  @ApiParam({
+    name: 'id',
+    description: 'Идентификатор категории',
+    required: true,
+  })
+  @Public()
   @Get(':id')
   async findOne(@Param('id') id: number): Promise<Category> {
     return await this.categoriesService.findOneById(id);
@@ -84,8 +109,17 @@ export class CategoriesController {
     description: 'Категория удалена',
     type: Category,
   })
+  @ApiBearerAuth()
+  @UseGuards(AuthGuard('jwt'))
   @Delete(':id')
-  async delete(@Param('id') id: number): Promise<Category> {
-    return await this.categoriesService.deleteById(id);
+  async delete(
+    @CurrentUser() user: UserPayload,
+    @Param('id') id: number,
+  ): Promise<Category> {
+    if (isAdmin(user)) {
+      return await this.categoriesService.deleteById(id);
+    } else {
+      throw new AccessRightsException();
+    }
   }
 }
